booking pages updated

+++ api/routes/tickets.js
@@ -38,20 +38,24 @@ router.delete("/checkout", async (req, res) => {
   // console.log(req.body);
   try {
     const ticket = await Tickets.findOne({ vehicle: req.body.v_no });
-    !ticket && res.status(401).json("Wrong credentials!");
+    if(!ticket)  res.status(401).json("Wrong credentials!");
+    else{
 
-    // console.log(ticket);
-    const hashedPassword = CryptoJS.AES.decrypt(
-      ticket.reff_no,
-      process.env.PASS_SEC
-    );
-    const OriginalPassword = hashedPassword.toString(CryptoJS.enc.Utf8);
-    OriginalPassword !== req.body.seckey && res.status(401).json("Wrong credentials!");
-    // console.log(OriginalPassword);  
-    
-    const delticket = await Tickets.findOneAndDelete({vehicle:req.body.v_no, reff_no:ticket.reff_no});
-    res.status(200).json(delticket);  
-    
+      // console.log(ticket);
+      const hashedPassword = CryptoJS.AES.decrypt(
+        ticket.reff_no,
+        process.env.PASS_SEC
+      );
+      const OriginalPassword = hashedPassword.toString(CryptoJS.enc.Utf8);
+      if(OriginalPassword !== req.body.seckey)  res.status(401).json("Wrong credentials!");
+      // console.log(OriginalPassword);  
+      else{
+
+        const delticket = await Tickets.findOneAndDelete({vehicle:req.body.v_no, reff_no:ticket.reff_no});
+        res.status(200).json(delticket);  
+      }
+      
+    }
   } catch (err) {
     res.status(500).json(err);
   }
